fix: revert bad g_vtxIn_x collapse for 3 functions at 0x007af958
The big batch in c22424d3 incorrectly collapsed two distinct address
slots into one symbol:
- 0x007af958: the start of the per-vertex array (was g_data_007af958)
- 0x007af95c: a single s16 value (canonical g_vtxIn_x per render.h)

Helper_EmitLine_004b2d20, AppInit_Misc3_004b2ac0, and
ProjectVertex_004b2e80 each encoded one of these two addresses in the
orig EXE. After the collapse, all three resolved to g_vtxIn_x via the
shared symbol map - but render.h pins g_vtxIn_x at 0x007af95c, so the
two functions that wanted 0x007af958 (Helper_EmitLine reading the
vertex array base, AppInit_Misc3 memset'ing it) produced bytes
shifted by 4.

Fix:
- Revert the prior /* 0x007af958 */ header-comment "fix" - the original
  /* 0x007af95c */ was correct (g_vtxIn_x truly lives at 0x007af95c
  matching the +0x6 stride described in the header)
- Restore g_data_007af958 as a distinct symbol in extras_map.yaml +
  src/data.c (it's the per-vertex array base, not the same slot as
  g_vtxIn_x)
- In src/boot/m_stack_boot_push4_init.c (Helper_EmitLine) and
  src/boot/app_init_misc3.c, rewrite the 4 references from g_vtxIn_x
  back to g_data_007af958

Brings synth Mismatches from 31 -> 26 (Helper_EmitLine, AppInit_Misc3,
and ProjectVertex now all synthesize byte-perfect). make matching
confirms MD5-identical. The 26 remaining mismatches are similar
collapse bugs in other addresses (0x7af9f7 / 0x7af9f3 RGB scale
cluster, 0x7af914 vs 0xf9f714 dual-section aliases, 0x7af950 vs
0x7af954, 0x7af938 vs 0x7af93c) plus the same s16-vs-int codegen-drift
pattern for the renderer/blit family. Tracked for follow-up.

+++ src/engine/copy_args16_to_globals.c
@@ -39,9 +39,9 @@ void AppShutdown(void) {
  *   load two dword globals, then move 16-bit args at [esp+4],[esp+8],[esp+0xc]
  *   into 16-bit globals; flush the dwords back too.
  */
-extern s16 g_vtxScreenX;
+extern unsigned int g_vtxScreenX;
 extern s32 g_vtxOut_z;
-extern s16 g_vtxScreenP2X;
+extern unsigned int g_vtxScreenP2X;
 extern s32 g_vtxOut2_z;
 extern unsigned short g_word_007af95c;
 extern unsigned short g_word_007af962;
